Make few entities and their jpa repository with findAll method and RestControlelr for each.
Entity in databse(hibernate also) mjesto,ustanova,vrstaUstanove,speciajlziacijaUstanove,
odjel,nazivOdjel, their jpaRepositories and restControllers with get only methods. All tested
in Angular with ajax calls and all works

+++ src/main/java/hr/tvz/baric/zavrsni/repo/ExampleJpaRepo.java
@@ -6,7 +6,7 @@ import hr.tvz.baric.zavrsni.model.Example;
 
 public interface ExampleJpaRepo extends JpaRepository<Example, Long> {
 	
-	public Example findById(Integer id);
+	public Example findById(Long id);
 	public Example findByIme(String ime);
 
 }
